refactor: moved some functions
