+++ inline_snapshot/_sentinels.py
@@ -1,0 +1,6 @@
+# sentinels
+class Undefined:
+    pass
+
+
+undefined = Undefined()
